refactor: remove circular reference
