+++ src/modules/real-time-metrics/reports/index.js
@@ -1,3 +1,4 @@
 import LoadReportsDataBySelectedDashboard from './load-reports-data-by-selected-dashboard'
+import ResolveReport from './resolve-reports'
 
-export { LoadReportsDataBySelectedDashboard }
+export { LoadReportsDataBySelectedDashboard, ResolveReport }
